fix cast

+++ android/plugin/src/main/java/com/tenforwardconsulting/bgloc/DistanceFilterLocationProvider.java
@@ -324,7 +324,7 @@ public class DistanceFilterLocationProvider extends AbstractLocationProvider imp
     }
 
     private Integer calculateDistanceFilter(Float speed) {
-		speed = 10.0;
+		speed = (float) 10;
         Double newDistanceFilter = (double) config.getDistanceFilter();
         if (speed < 100) {
             float roundedDistanceFilter = (round(speed / 5) * 5);
